feat: added proper bin file
Also a little bit of implementation towards connecting bin command to actual analyser
