+++ .eslintrc.json
@@ -12,7 +12,7 @@
   "rules": {
     "accessor-pairs": "error",
     "array-bracket-newline": "error",
-    "array-bracket-spacing": "error",
+    "array-bracket-spacing": ["error", "always"],
     "array-callback-return": "error",
     "array-element-newline": "error",
     "arrow-body-style": "error",
@@ -218,7 +218,7 @@
     "no-with": "error",
     "nonblock-statement-body-position": "error",
     "object-curly-newline": "error",
-    "object-curly-spacing": "error",
+    "object-curly-spacing": ["error", "always"],
     "object-property-newline": "error",
     "object-shorthand": "error",
     "one-var": "off",
