+++ package.json
@@ -1,17 +1,20 @@
 {
+  "bin": {
+    "regana": "./bin/regana.js"
+  },
   "engines": {
     "node": ">=8"
   },
   "name": "regana",
   "version": "0.1.0",
   "description": "Tool to analyse risk for regressions based on git diff",
-  "main": "index.js",
+  "main": "src/regana.js",
   "scripts": {
     "docs": "node ./node_modules/.bin/jsdoc --configure .jsdocrc.json",
     "build": "node console.log('build');",
     "start": "node console.log('start');",
     "test": "node ./node_modules/.bin/jest --config=.jestrc.json",
-    "lint": "node ./node_modules/.bin/eslint --config=.eslintrc.json src/**/**",
+    "lint": "node ./node_modules/.bin/eslint --config=.eslintrc.json src/**/** bin/**/**",
     "codecov": "node ./node_modules/.bin/codecov",
     "commitmsg": "commitlint -e $GIT_PARAMS",
     "precommit": "npm run test && npm run lint && npm run docs"
@@ -38,6 +41,7 @@
     "@commitlint/cli": "^6.2.0",
     "@commitlint/config-conventional": "^6.1.3",
     "@commitlint/travis-cli": "^6.2.0",
+    "chalk": "^2.4.1",
     "codecov": "^3.0.1",
     "eslint": "^4.19.1",
     "eslint-plugin-jest": "^21.15.1",
