+++ src/regana.js
@@ -1,0 +1,7 @@
+
+/**
+ * @file Expose out Analyser and other tools
+ * @author Matti Mehtonen
+ */
+
+module.exports = {};
